Improve Windows build
Libraries may be more difficult to setup on Windows. This changes
add the possibility to disable testing (and Boost includes) using
CMakeCache.

The SFML_ROOT variable is now also cached in order to be modified more
easily. Policy CMP0074 has to be enabled in order to all find_package
to use <PackageName>_ROOT variable to help find SFML.