Fix include directories for libraries
All the libraries used to include the base source directory of the
library as INTERFACE. This is ok for importing the library and using
the headers defined inside, but this configuration does not add the
source directory of the library as include directory when compiling the
library itself. Changing the scope from INTERFACE to PUBLIC allows to
use the root directory as include source for both the compilation of
the library and the compilation of code using this library.